+++ vite.config.mjs
@@ -84,7 +84,7 @@ export default defineConfig(({ mode }) => ({
       : {},
   test: {
     include: ['test/unit/**/*.{js,jsx}', 'test/jest/**/*.test.{js,jsx}'],
-    poolOptions: { threads: { singleThread: true } },
+    poolOptions: { forks: { singleFork: true } },
     globals: true,
     clearMocks: true,
     environment: 'jsdom',