+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.134.0] - Not released
+### Added
+- Add support for inline code (single backticks) and code blocks (triple backticks)
+  to post/comment text parser.
+### Fixed
+- Autocomplete menu did not update properly after the server autocomplete
+  response.
+
 ## [1.133.0] - 2024-07-04
 ### Added
 - Add a "description" meta-tag to the index.html.